pyproject.toml
[project]
name = "llm-onboarding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai-agents>=0.3.3",
    "pydantic[email]>=2.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]
